refactor: rename StreamingResultSet to DataCloudResultSet, drop marker interface
Pre-unify there were three result-set implementations: StreamingResultSet
(streaming Arrow query results), DataCloudMetadataResultSet (metadata),
SimpleResultSet (in-memory rows). The DataCloudResultSet interface — a
one-method (getQueryId) extension over java.sql.ResultSet — was the
common "implements" the public API surfaced; StreamingResultSet was the
only one that ever implemented it as a non-trivial impl.

The unify refactor collapsed all three implementations into
StreamingResultSet, but kept the interface and the "Streaming" name. Two
problems fall out:

- The "Streaming" name now lies. Metadata results flow through the same
  class but they're a one-shot in-memory IPC blob — nothing streaming
  about them. MetadataResultSets.of even passes /*queryId=*/ null
  because there is no query.
- The DataCloudResultSet interface has one implementer and one method.
  Layering an interface for one impl is just a reader trap: callers
  instinctively look for "what other implementations exist" and find
  none.

Collapse the two:
- Rename the class StreamingResultSet -> DataCloudResultSet.
- Delete the old DataCloudResultSet interface (the public method
  getQueryId() now lives directly on the class via @Getter).
- Update all production and test references; rename the affected test
  files to match (StreamingResultSet*Test -> DataCloudResultSet*Test).

The public API surface is unchanged in source for the common cases:
DataCloudConnection.getRowBasedResultSet / getChunkBasedResultSet still
return DataCloudResultSet, just as a class instead of an interface.
This is binary-incompatible for any caller that ever cast to or
implemented the old interface; in practice only StreamingResultSet
implemented it on the read side, and no code outside the driver
implemented it on the write side.
